+++ spec/bibliothecary_spec.rb
@@ -37,6 +37,7 @@ describe Bibliothecary do
           Bibliothecary::Parsers::Meteor,
           Bibliothecary::Parsers::MLflow,
           Bibliothecary::Parsers::Nimble,
+          Bibliothecary::Parsers::Nix,
           Bibliothecary::Parsers::NPM,
           Bibliothecary::Parsers::Nuget,
           Bibliothecary::Parsers::Ollama,